Add files via upload
Throughout the semester, I gained significant experience in MATLAB through various projects completed with my colleagues. These projects not only enhanced my technical skills but also taught me the importance of teamwork and collaboration. It was an invaluable experience that greatly contributed to my development. Here are the projects I worked on with my team